requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
tqdm==4.66.5
PyYAML==6.0.2
rich==13.8.1

# Indexing / ANN
faiss-cpu==1.8.0.post1

# Vision & CAM utilities
open-clip-torch==2.26.1
timm==0.9.16
einops==0.8.0
opencv-python-headless==4.10.0.84
imageio==2.35.1
imageio-ffmpeg==0.4.9
matplotlib==3.9.2

# Medical I/O (optional but useful)
pydicom==2.4.4
nibabel==5.3.1

# Simple probes / eval
scikit-learn==1.5.2

# (Optional) HF models for BiomedCLIP or text experiments
transformers==4.45.2
sentencepiece==0.2.0
